requirements.txt
Flask
Flask-Limiter
Flask-Compress
Flask-Caching
python-dotenv
sqlcipher3
redis
werkzeug
